Rate limiting added for cleaning lists and infinite loop bug fixed
Bug report #36 identified the need to rate limit list importing and this was fixed by @andrewchenk in #39 by adding a 1-second sleep to api_add_to_list(). This commit adds the same fix for cleaning lists (-C option), specifically adding the same sleep to api_remove_from_list().

Fixed an infinite loop that occurs if the -C option is used when the target list is already empty.